feat: 역 삭제 시 예외 처리 추가 - exception 수정

+++ src/main/java/wooteco/subway/service/StationService.java
@@ -3,20 +3,25 @@ package wooteco.subway.service;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
+import wooteco.subway.domain.section.Sections;
 import wooteco.subway.domain.station.Station;
 import wooteco.subway.dto.StationRequest;
 import wooteco.subway.dto.StationResponse;
+import wooteco.subway.exception.NotRemovableStationException;
 import wooteco.subway.exception.StationDuplicationException;
 import wooteco.subway.exception.StationNotFoundException;
+import wooteco.subway.repository.SectionDao;
 import wooteco.subway.repository.StationDao;
 
 @Service
 public class StationService {
 
     private final StationDao stationDao;
+    private final SectionDao sectionDao;
 
-    public StationService(StationDao stationDao) {
+    public StationService(StationDao stationDao, SectionDao sectionDao) {
         this.stationDao = stationDao;
+        this.sectionDao = sectionDao;
     }
 
     public StationResponse createStation(StationRequest stationRequest) {
@@ -42,6 +47,9 @@ public class StationService {
     }
 
     public void deleteById(Long id) {
+        if (sectionDao.findByStation(id) > 0) {
+            throw new NotRemovableStationException();
+        }
         stationDao.deleteById(id);
     }
